CRUD/requirements.txt
mysql-connector-python==9.0.0
prettytable==3.10.0
python-dotenv==1.0.1
